requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1

# MongoDB stack — these three must be compatible with each other.
# Tested working on Python 3.13:
#   pymongo 4.9.x removed _QUERY_OPTIONS which motor 3.5.x depended on.
#   motor 3.6.0 + pymongo 4.9.2 + beanie 1.27.0 resolves this cleanly.
beanie==1.27.0
motor==3.6.0
pymongo==4.9.2

# Zerodha
kiteconnect==5.0.1

# HTTP client
httpx==0.27.0

# Retry logic
tenacity==8.3.0

# Date parsing
python-dateutil==2.9.0

# Background jobs
celery==5.4.0
redis==5.0.8

# PDF parsing
# casparser and pdfplumber both work with pdfminer-six==20240706 at runtime.
# Install pdfplumber with --no-deps to skip its overly strict pdfminer-six pin.
# See: pip install pdfplumber --no-deps && pip install Pillow pypdfium2
pymupdf  # required for some CAS PDF types (GPL license) chardet
casparser==0.8.1
pdfminer-six==20240706
pdfplumber==0.11.0   # install with --no-deps
Pillow==10.4.0
pypdfium2
pymupdf  # required for some CAS PDF types (GPL license)
chardet
pytesseract==0.3.10

# Market data
# yfinance removed — using NSE direct API instead

# Embeddings
voyageai==0.2.3

# Linting / dev
ruff==0.5.0

python-multipart

scipy>=1.13.0
numpy>=1.26.0